Add container control examples
Add registery scanner examples
Update Java search examples to use Double for ID values

+++ java/src/com/trendmicro/deepsecurity/docs/ApplicationControlExamples.java
@@ -82,7 +82,7 @@ public class ApplicationControlExamples {
 		SearchCriteria computerCriteria = new SearchCriteria();
 		computerCriteria.setFieldName("computerID");
 		computerCriteria.setNumericTest(NumericTestEnum.EQUAL);
-		computerCriteria.setNumericValue(Long.valueOf(computerId.longValue()));
+		computerCriteria.setNumericValue(computerId.doubleValue());
 
 		// Search filter
 		SearchFilter searchFilter = new SearchFilter();
